fix: rename caps.py to capabilities.py
