+++ ex_app/lib/capabilities.py
